Adding test cases for queries that containt keywords
These are test cases for the frontend of YSmart. Non of the test cases pass at this point with the frontend generated with the original YSmart grammar. I need to update the grammar and generate the front end again but it means that the new frontent may break the backend.


git-svn-id: http://ysmart.googlecode.com/svn/branches/pure-python@254 1d6e5d1b-7b3d-7f63-4173-3946eeb354c0