Fixing the bug in naming generated MapReduce jobs.
This change is to make YSmart consistent with the online version.
The online version starts file names with a basename + a number. The number starts at 1.
E.g.: TestQuery1.java, TestQuery2.java, ...

This new version now does the same with one difference: It pads the number with zeros to make it 3 digit long.
E.g.: TestQuery001.java TestQuery002.java



git-svn-id: http://ysmart.googlecode.com/svn/branches/development@279 1d6e5d1b-7b3d-7f63-4173-3946eeb354c0